fix:review_count 필드 float로 내려오는 이슈 수정

+++ app/schema/bakery.py
@@ -11,7 +11,7 @@ class RecommendBakery(BaseModel):
     bakery_id: int = Field(..., description="베이커리 id")
     bakery_name: str = Field(..., description="베이커리 상호명")
     avg_rating: float = Field(..., description="평균별점")
-    review_count: float = Field(..., description="리뷰 개수")
+    review_count: int = Field(..., description="리뷰 개수")
     is_opened: bool = Field(..., description="영업 여부")
     img_url: str = Field(..., description="베이커리 썸네일")
